fix cli
